RECALLOGUE/requirements.txt
FastAPI
pydantic
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
asyncpg
python-dotenv
uvicorn
python-multipart
google-generativeai
requests
chromadb
pdfplumber
python-docx
chardet
pytest